+++ Chips/Chip.js
@@ -5,8 +5,9 @@ import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 import {use, useCallback, useEffect, useRef} from 'react';
-
 import Button from '../Button';
+import Icon from '../Icon';
+import Image from '../Image';
 import $L from '../internal/$L';
 import Skinnable from '../Skinnable';
 
@@ -31,7 +32,8 @@ const chipDeleteButtonShape = PropTypes.shape({
 });
 
 const ChipDefaultProps = {
-	disabled: false
+	disabled: false,
+	imageSize: 24
 };
 
 /**
@@ -58,11 +60,11 @@ const ChipDefaultProps = {
 const ChipBase = (props) => {
 	const {handleChipDelete, getNextTargetFromDeleteButton, registerChild} = use(ChipsContext);
 	const chipProps = setDefaultProps(props, ChipDefaultProps);
-	const {children, className, deleteButton, disabled, icon, id, onClick, ref, ...rest} = chipProps;
+	const {checked, children, className, deleteButton, disabled, icon, id, imageSize, isImage, onClick, ref, ...rest} = chipProps;
 
-	const chipClassName = classnames(className, deleteButton?.position);
+	const ariaLabel = children + ' ' + $L('chip') + (checked ? $L('checked') : $L('unchecked'));
 	const buttonClassName = classnames(css.deleteButtonContainer, css[deleteButton?.position || 'right']);
-
+	const chipClassName = classnames(className, deleteButton?.position);
 	const containerRef = useRef(null);
 	const clientRef = useRef(null);
 	const deleteButtonRef = useRef(null);
@@ -150,6 +152,14 @@ const ChipBase = (props) => {
 		}
 	}, [deleteButton, handleChipDelete, id]);
 
+	const iconComponent = ({children: icon, ...props}) => {
+		return <>
+			{checked && <Icon {...props} children={'check'} />}
+			{isImage && <Image {...props} style={{borderRadius: '999px', width: `${imageSize}px`, height: `${imageSize}px`}} src={icon} />}
+			{!isImage && <Icon {...props}>{icon}</Icon>}
+		</>
+	}
+
 	return (
 		<div
 			{...rest}
@@ -162,13 +172,14 @@ const ChipBase = (props) => {
 			ref={containerRef}
 		>
 			<Button
-				aria-label={children + ' ' + $L("Chip")}
+				aria-label={ariaLabel}
 				css={css}
 				className={chipClassName}
 				data-chip-index={id}
 				disabled={disabled}
 				focusEffect="static"
 				icon={icon ? icon : null}
+				iconComponent={iconComponent}
 				size="small"
 				onFocus={handleFocus}
 				onClick={onClick}
@@ -200,6 +211,14 @@ const ChipBase = (props) => {
 ChipBase.displayName = 'Chip';
 
 ChipBase.propTypes = /** @lends limestone/Chips.Chip.prototype */ {
+	/**
+	 * Sets the chip as `checked` if `true`
+	 *
+	 * @type {Boolean}
+	 * @public
+	 */
+	checked: PropTypes.bool,
+
 	/**
 	 * A label displayed in the chip content.
 	 *
@@ -245,7 +264,23 @@ ChipBase.propTypes = /** @lends limestone/Chips.Chip.prototype */ {
 	 * @type {String|Object}
 	 * @public
 	 */
-	icon: PropTypes.oneOfType([PropTypes.string, PropTypes.object])
+	icon: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+
+	/**
+	 * Sets the size of the image passed to the component.
+	 *
+	 * @type {Number}
+	 * @public
+	 */
+	imageSize: PropTypes.number,
+
+	/**
+	 * If `true` tells the component to use `Image` instead of `Icon` element inside `Chip`.
+	 *
+	 * @type {Boolean}
+	 * @public
+	 */
+	isImage: PropTypes.bool
 };
 
 /**