+++ Chips/Chips.js
@@ -1,4 +1,5 @@
 import {setDefaultProps} from '@enact/core/util';
+import IString from 'ilib/lib/IString';
 import Spotlight from '@enact/spotlight';
 import {SpotlightContainerDecorator} from '@enact/spotlight/SpotlightContainerDecorator';
 import {getAllContainerIds} from '@enact/spotlight/src/container';
@@ -7,6 +8,7 @@ import classnames from 'classnames';
 import PropTypes from 'prop-types';
 import compose from 'ramda/src/compose';
 import {createContext, useCallback, useRef} from 'react';
+import $L from '../internal/$L';
 
 import css from './Chips.module.less';
 
@@ -43,6 +45,8 @@ const ChipsBase = (props) => {
 	const chipsClassName = classnames(css.chips, css[orientation], className);
 	const childRefs = useRef([]);
 	const containerRef = useRef(null);
+	const ariaLabel = new IString($L('{total} items in total')).format({total: children.length});
+	const id = Math.random().toString(36).substring(2, 10);
 
 	const getPreviousChip = useCallback((id) => {
 		const currentIndex = childRefs.current.findIndex((child) => child.id === id);
@@ -120,16 +124,24 @@ const ChipsBase = (props) => {
 	}, [children]);
 
 	return (
-		<div {...rest} className={chipsClassName} ref={containerRef}>
-			<ChipsContext
-				value={{
-					getNextTargetFromDeleteButton,
-					handleChipDelete,
-					registerChild
-				}}
+		<div  role="region" aria-labelledby={`${id}_chips`}>
+			<div id={`${id}_chips`}
+				 role='group'
+				 aria-label={ariaLabel}
+				 className={chipsClassName}
+				 ref={containerRef}
+				 {...rest}
 			>
-				{children}
-			</ChipsContext>
+				<ChipsContext
+					value={{
+						getNextTargetFromDeleteButton,
+						handleChipDelete,
+						registerChild
+					}}
+				>
+					{children}
+				</ChipsContext>
+			</div>
 		</div>
 	);
 };